requirements.txt
pandas
numpy
matplotlib
scikit-learn
pyarrow
pyyaml
fastapi
uvicorn
jinja2
fredapi
hmmlearn
